Xft: Change xftfont_min_descent to return max descent
Like the font's ascent, its descent is also always positive.  That means
for the common use-case of getting the greatest descent for a list of
fonts, one can simply return the maximum of all descents.  In this way,
it is possible to center text that's using multiple fonts by simply
calculating the maximum ascent and descent.  This is also done in
xmobar's MinXft library[1], which this part of the code is modeled
after.

Fixes: https://github.com/xmonad/X11-xft/commit/0cc513a9e36e8ea0abe97816fb69acb0af484f44
Related: https://github.com/xmonad/xmonad-contrib/pull/646
[1]: https://github.com/jaor/xmobar/blob/master/src/Xmobar/X11/MinXft.hsc#L94